+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Change `ignore_changes` of lifecycle for EC2 instance: use hard coded values `ami` and `user_data` ([#3](https://github.com/WildBeavers/terraform-aws-ec2-instance/issues/3))
 - Support new Terraform 0.12 feature: EC2 instance as output  ([#4](https://github.com/WildBeavers/terraform-aws-ec2-instance/issues/4))
 - The format string for generating the hostname of an instance can be overridden ([#9](https://github.com/WildBeavers/terraform-aws-ec2-instance/issues/9))
+- support suffix for volume tag `Name`
 
 <a name="v2.6.0"></a>
 ## [v2.6.0] - 2019-07-21